refactor: consolidate dashboard creation and remove deprecated scripts
